// src/api/generated/models/getRequirementsIdLinkChangesParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Requirements Backend API
 * HTTP API for the Requirements Management app.
 * OpenAPI spec version: 0.0.1
 */

export type GetRequirementsIdLinkChangesParams = {
/**
 * @minimum 1
 */
fromRevision: number;
/**
 * @minimum 1
 */
toRevision: number;
};