// src/api/generated/models/requirementLinkChangesResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Requirements Backend API
 * HTTP API for the Requirements Management app.
 * OpenAPI spec version: 0.0.1
 */
import type { RequirementLinkResponseDto } from './requirementLinkResponseDto';

export interface RequirementLinkChangesResponseDto {
  requirementId: string;
  /** @minimum 1 */
  fromRevision: number;
  /** @minimum 1 */
  toRevision: number;
  addedOutgoingLinks: RequirementLinkResponseDto[];
  removedOutgoingLinks: RequirementLinkResponseDto[];
  unchangedOutgoingLinks: RequirementLinkResponseDto[];
  addedIncomingLinks: RequirementLinkResponseDto[];
  removedIncomingLinks: RequirementLinkResponseDto[];
  unchangedIncomingLinks: RequirementLinkResponseDto[];
}