// src/api/generated/models/requirementLinkHistoryResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Requirements Backend API
 * HTTP API for the Requirements Management app.
 * OpenAPI spec version: 0.0.1
 */
import type { RequirementLinkHistoryResponseDtoEventType } from './requirementLinkHistoryResponseDtoEventType';
import type { RequirementLinkHistoryResponseDtoRelationshipType } from './requirementLinkHistoryResponseDtoRelationshipType';

export interface RequirementLinkHistoryResponseDto {
  id: string;
  linkId: string;
  projectId: string;
  eventType: RequirementLinkHistoryResponseDtoEventType;
  relationshipType: RequirementLinkHistoryResponseDtoRelationshipType;
  sourceRequirementId: string;
  sourceVisibleKey: string;
  /** @nullable */
  oldTargetRequirementId: string | null;
  /** @nullable */
  oldTargetVisibleKey: string | null;
  /** @nullable */
  newTargetRequirementId: string | null;
  /** @nullable */
  newTargetVisibleKey: string | null;
  occurredAt: string;
  /** @nullable */
  actor: string | null;
  /** @nullable */
  reason: string | null;
}