// src/api/generated/models/requirementLinkResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Requirements Backend API
 * HTTP API for the Requirements Management app.
 * OpenAPI spec version: 0.0.1
 */
import type { RequirementLinkResponseDtoRelationshipType } from './requirementLinkResponseDtoRelationshipType';
import type { RequirementStatus } from './requirementStatus';
import type { RequirementType } from './requirementType';

export interface RequirementLinkResponseDto {
  id: string;
  projectId: string;
  relationshipType: RequirementLinkResponseDtoRelationshipType;
  sourceRequirementId: string;
  sourceVisibleKey: string;
  sourceType: RequirementType;
  sourceCategoryId: string;
  sourceCategoryKey?: string;
  sourceStatus: RequirementStatus;
  targetRequirementId: string;
  targetVisibleKey: string;
  targetType: RequirementType;
  targetCategoryId: string;
  targetCategoryKey?: string;
  targetStatus: RequirementStatus;
  createdAt: string;
  updatedAt: string;
}