// src/api/generated/models/requirementRevisionLinksResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Requirements Backend API
 * HTTP API for the Requirements Management app.
 * OpenAPI spec version: 0.0.1
 */
import type { RequirementLinkResponseDto } from './requirementLinkResponseDto';

export interface RequirementRevisionLinksResponseDto {
  requirementId: string;
  /** @minimum 1 */
  revisionNumber: number;
  revisionCreatedAt: string;
  outgoingLinks: RequirementLinkResponseDto[];
  incomingLinks: RequirementLinkResponseDto[];
}